@@ -7,46 +7,65 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "generate": "graphql-codegen"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.1.1",
-    "@angular/common": "~8.1.1",
-    "@angular/compiler": "^8.2.1",
-    "@angular/core": "^8.2.1",
-    "@angular/forms": "^8.2.1",
-    "@angular/platform-browser": "^8.2.1",
-    "@angular/platform-browser-dynamic": "^8.2.1",
-    "@angular/router": "^8.2.1",
-    "angular-user-idle": "^2.1.4",
+    "@angular/animations": "^10.0.14",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^10.0.14",
+    "@angular/compiler": "^10.0.14",
+    "@angular/core": "^10.0.14",
+    "@angular/forms": "^10.0.14",
+    "@angular/platform-browser": "^10.0.14",
+    "@angular/platform-browser-dynamic": "^10.0.14",
+    "@angular/router": "^10.0.14",
+    "@graphql-codegen/near-operation-file-preset": "^1.17.11",
+    "angular-user-idle": "^2.2.4",
+    "apollo-angular": "^1.10.0",
+    "apollo-angular-link-http": "^1.11.0",
+    "apollo-cache-inmemory": "^1.6.6",
+    "apollo-client": "^2.6.10",
     "chart.js": "^2.8.0",
     "core-js": "^2.5.4",
-    "ng2-charts": "^2.3.0",
-    "rxjs": "~6.5.2",
+    "devextreme": "20.1.7",
+    "devextreme-angular": "20.1.7",
+    "dotenv": "^8.2.0",
+    "graphql": "^14.7.0",
+    "graphql-tag": "^2.11.0",
+    "install": "0.13.0",
+    "lodash-es": "^4.17.15",
+    "mysql": "^2.18.1",
+    "ngx-avatar": "^3.7.0",
+    "rxjs": "^6.6.2",
     "sweetalert2": "^8.15.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.13.0",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.801.3",
-    "@angular/cli": "^8.2.1",
-    "@angular/compiler-cli": "^8.2.1",
-    "@angular/language-service": "^8.2.1",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.1",
+    "@angular-devkit/build-angular": "^0.1000.8",
+    "@angular/cli": "^10.0.8",
+    "@angular/compiler-cli": "^10.0.14",
+    "@angular/language-service": "^10.0.14",
+    "@graphql-codegen/cli": "^1.17.8",
+    "@graphql-codegen/typescript": "^1.17.9",
+    "@graphql-codegen/typescript-apollo-angular": "^1.17.7",
+    "@graphql-codegen/typescript-operations": "^1.17.8",
+    "@types/jasmine": "^2.8.17",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.12.54",
+    "codelyzer": "^5.2.2",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
+    "karma": "^5.1.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "protractor": "^7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.4.5"
+    "typescript": "~3.9.7"
   }
 }
