@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import Swal from 'sweetalert2';
 import { UsuarioService } from '../services/services.index';
-import { Usuario } from '../models/usuario.model';
 import { Router } from '@angular/router';
+import { RegisterData } from './login.interface';
 
 declare function init_plugins();
 
@@ -16,6 +16,13 @@ declare function init_plugins();
 export class RegisterComponent implements OnInit {
 
   forma: FormGroup;
+
+  register: RegisterData = {
+    nombre: '',
+    email: '',
+    password: ''
+
+  };
 
   constructor(
 // tslint:disable-next-line: variable-name
@@ -47,13 +54,12 @@ export class RegisterComponent implements OnInit {
     init_plugins();
 
     this.forma = new FormGroup({
-      nombre: new FormControl( 'null', Validators.required ),
-      correo: new FormControl('null', [Validators.required, Validators.email] ),
-      password: new FormControl('null', Validators.required ),
-      password2: new FormControl('null', Validators.required ),
+      nombre: new FormControl( '', Validators.required ),
+      correo: new FormControl('', [Validators.required, Validators.email] ),
+      password: new FormControl('', Validators.required ),
+      password2: new FormControl('', Validators.required ),
       condiciones: new FormControl( false )
     }, { validators: this.sonIguales( 'password', 'password2') });
-
 
   }
 
@@ -74,15 +80,20 @@ export class RegisterComponent implements OnInit {
       return;
     }
 
-    const usuario = new Usuario (
-        this.forma.value.nombre,
-        this.forma.value.correo,
-        this.forma.value.password
-      );
+    this.register.nombre = this.forma.value.nombre;
+    this.register.email = this.forma.value.correo;
+    this.register.password = this.forma.value.password;
 
-    this._usuarioService.crearUsuario( usuario )
-      .subscribe(resp => this.router.navigate(['/login']));
+    this._usuarioService.register( this.register )
+     // .subscribe( (resp: any)  =>   this.router.navigate(['/login']));
+     .subscribe(() => {
+      this.router.navigate(['/login']);
+    }, (error) => {
+      console.log('there was an error sending the query', error);
+    });
 
-    }
+
+
+  }
 
 }
