@@ -1,17 +1,12 @@
 import { NgModule } from '@angular/core';
 import { PAGES_ROUTES } from './pages.routes';
 
-// ng2-charts
-import { ChartsModule } from 'ng2-charts';
-
 
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { ProgressComponent } from './progress/progress.component';
-import { Graficas1Component } from './graficas1/graficas1.component';
 import { PagesComponent } from './pages.component';
 import { SharedModule } from '../shared/shared.module';
-import { FormsModule } from '@angular/forms';
-import { GraficoDonaComponent } from '../components/grafico-dona/grafico-dona.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AccountSettingsComponent } from './account-settings/account-settings.component';
 import { PromesasComponent } from './promesas/promesas.component';
 import { RxjsComponent } from './rxjs/rxjs.component';
@@ -25,9 +20,13 @@ import { MedicosComponent } from './medicos/medicos.component';
 import { MedicoComponent } from './medicos/medico.component';
 import { BusquedaComponent } from './busqueda/busqueda.component';
 import { UserIdleModule } from 'angular-user-idle';
-
-
-
+import { DxButtonModule, DxDataGridModule, DxToolbarModule, DxPopupModule, DxBoxModule,
+         DxSelectBoxModule, DxTextBoxModule, DxValidatorModule } from 'devextreme-angular';
+import { BrowserModule } from '@angular/platform-browser';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import { UsuarioComponent } from './usuarios/usuario.component';
+import { ComponentsModule } from '../components/components.module';
+import { AvatarModule } from 'ngx-avatar';
 
 
 
@@ -39,8 +38,6 @@ import { UserIdleModule } from 'angular-user-idle';
         DashboardComponent,
         ProgressComponent,
         PagesComponent,
-        Graficas1Component,
-        GraficoDonaComponent,
         AccountSettingsComponent,
         PromesasComponent,
         RxjsComponent,
@@ -51,26 +48,39 @@ import { UserIdleModule } from 'angular-user-idle';
         MedicosComponent,
         MedicoComponent,
         BusquedaComponent,
-        
+        UsuarioComponent,
 
 
     ],
 
     exports: [
         DashboardComponent,
-        ProgressComponent,
-        Graficas1Component,
+        ProgressComponent
+
     ],
     imports: [
         SharedModule,
         PAGES_ROUTES,
         FormsModule,
-        ChartsModule,
         PipesModule,
         CommonModule,
-        UserIdleModule.forRoot({idle: 6, timeout: 300, ping: 120})
-
-    ]
+        UserIdleModule.forRoot({idle: 6, timeout: 300, ping: 120}),
+        DxButtonModule,
+        DxDataGridModule,
+        BrowserModule,
+        DxToolbarModule,
+        ComponentsModule,
+        ReactiveFormsModule,
+        BrowserAnimationsModule,
+        AvatarModule,
+        DxPopupModule,
+        DxBoxModule,
+        DxSelectBoxModule,
+        DxTextBoxModule,
+        DxButtonModule,
+        DxValidatorModule
+        
+    ],
 
 
 })
