@@ -2,7 +2,6 @@ import { RouterModule, Routes, CanActivate } from '@angular/router';
 import { PagesComponent } from './pages.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { ProgressComponent } from './progress/progress.component';
-import { Graficas1Component } from './graficas1/graficas1.component';
 import { AccountSettingsComponent } from './account-settings/account-settings.component';
 import { PromesasComponent } from './promesas/promesas.component';
 import { RxjsComponent } from './rxjs/rxjs.component';
@@ -11,9 +10,11 @@ import { ProfileComponent } from './profile/profile.component';
 import { UsuariosComponent } from './usuarios/usuarios.component';
 import { HospitalesComponent } from './hospitales/hospitales.component';
 import { MedicosComponent } from './medicos/medicos.component';
-import { MedicoComponent } from './medicos/medico.component';
 import { BusquedaComponent } from './busqueda/busqueda.component';
 import { AdminGuard } from '../services/services.index';
+import { UsuarioComponent } from './usuarios/usuario.component';
+import { VerificaTokenGuard } from '../services/guards/verifica-token.guard';
+
 
 
 
@@ -25,7 +26,6 @@ const pagesRoutes: Routes = [
         children: [
             { path: 'dashboard', component: DashboardComponent, data: { titulo: 'Dashboard'} },
             { path: 'progress', component: ProgressComponent, data: { titulo: 'Progress' } },
-            { path: 'graficas1', component: Graficas1Component, data: { titulo: 'Graficas' } },
             { path: 'promesas', component: PromesasComponent, data: { titulo: 'Promesas' } },
             { path: 'rxjs', component: RxjsComponent, data: { titulo: 'Observables' } },
             { path: 'account-setting', component: AccountSettingsComponent, data: { titulo: 'Ajustes del tema' }},
@@ -34,19 +34,29 @@ const pagesRoutes: Routes = [
 
 
             // Mantenimientos
-            { 
+            {
                 path: 'usuarios',
                 component: UsuariosComponent,
                 canActivate : [ AdminGuard ],
                 data: { titulo: 'Mantenimiento de usuarios' }
             },
-            { path: 'hospitales', component: HospitalesComponent, data: { titulo: 'Mantenimiento de hospitales' } },
+            {
+                path: 'hospitales',
+                component: HospitalesComponent,
+                canActivate : [ VerificaTokenGuard ],
+                data: { titulo: 'Mantenimiento de hospitales' }
+            },
             { path: 'medicos', component: MedicosComponent, data: { titulo: 'Mantenimiento de medicos' } },
-            { path: 'medico/:id', component: MedicoComponent, data: { titulo: 'Actualizar medico' } },
+            { path: 'usuario/:id', component: UsuarioComponent, data: { titulo: 'Actualizar usuario' } },
             { path: '', redirectTo: '/dashboard', pathMatch: 'full' },
+
+            // Elementos
+
+           // { path: 'grid', component: GridComponent, data: { titulo: '' } },
 
         ]
     },
+
 
 
 ];
