@@ -1,7 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { CanActivate } from '@angular/router';
-import { Observable } from 'rxjs';
 import { UsuarioService } from '../usuario/usuario.service';
 
 @Injectable({
@@ -12,14 +10,14 @@ export class AdminGuard implements CanActivate {
 
   constructor(
         public _usuarioService: UsuarioService
-       
+
   ) {
 
   }
 
   canActivate() {
 
-    if ( this._usuarioService.usuario.role === 'ADMIN_ROLE') {
+    if ( this._usuarioService.usuario.role === 'ADMIN') {
       return true;
     } else {
       this._usuarioService.logout();
