@@ -2,18 +2,24 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { ModalUploadService } from '../components/modal-upload/modal-upload.service';
+import { UsersValidator} from '../validators/usuarios/users-validators';
+
+
+
 
 
 import {
     SettingsService,
-    SharedService,
+    FormsService,
     SidebarService,
     UsuarioService,
+    ClienteService,
     LoginGuardGuard,
     SubirArchivoService,
     HospitalService,
     MedicoService,
-    AdminGuard
+    AdminGuard,
+    VerificaTokenGuard
 
 
  } from './services.index';
@@ -27,16 +33,20 @@ import {
   ],
   providers: [
     SettingsService,
-    SharedService,
+    FormsService,
     SidebarService,
     UsuarioService,
+    ClienteService,
     LoginGuardGuard,
     SubirArchivoService,
     ModalUploadService,
     HospitalService,
     MedicoService,
-    AdminGuard
+    AdminGuard,
+    UsersValidator,
+    VerificaTokenGuard
   ]
+  
 
 })
 export class ServiceModule { }
