@@ -1,15 +1,19 @@
+import { ClienteService } from './cliente/cliente.service';
+
 
 // Guards
 export { AdminGuard } from './guards/admin.guard';
 export { LoginGuardGuard } from './guards/login-guard.guard';
+export { VerificaTokenGuard } from './guards/verifica-token.guard';
 
 
 export { MedicoService } from './medico/medico.service';
 export { HospitalService } from './hospital/hospital.service';
 export { SubirArchivoService } from './subir-archivo/subir-archivo.service';
 export { UsuarioService } from './usuario/usuario.service';
+export { ClienteService} from './cliente/cliente.service';
 export { SettingsService } from './setting/settings.service';
-export { SharedService   } from './shared/shared.service';
+export { FormsService   } from './shared/forms.service';
 export { SidebarService } from './shared/sidebar.service';
 
 
