@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { UsuarioService } from 'src/app/services/services.index';
 import { Usuario } from '../../models/usuario.model';
 import { Router } from '@angular/router';
+import { User } from '../../../../.src/app/generated/types';
+
+//const path = require('path');
 
 @Component({
   selector: 'app-header',
@@ -10,13 +13,13 @@ import { Router } from '@angular/router';
 })
 export class HeaderComponent implements OnInit {
 
-  usuario: Usuario;
+  usuario: User;
 
-  constructor( public _usuarioService: UsuarioService,
+  constructor( public _USUARIOSERVICE: UsuarioService,
                public router: Router) { }
 
   ngOnInit() {
-    this.usuario = this._usuarioService.usuario;
+    this.usuario = this._USUARIOSERVICE.usuario;
   }
 
   buscar( termino: string ) {
